+++ lib/td_dd/search/mappings.ex
@@ -35,6 +35,14 @@ defmodule TdDd.Search.Mappings do
           external_id: %{type: "text", fields: @raw}
         }
       },
+      domain_parents: %{
+        type: "nested",
+        properties: %{
+          id: %{type: "long"},
+          name: %{type: "text", fields: @raw},
+          external_id: %{type: "text", fields: @raw}
+        }
+      },
       parent: %{
         properties: %{
           name: %{type: "text", fields: @raw},